Splits maliput_query_node_tests into multiple files per service.
- Moves specifc service request tests to individual files.
- Creates a function to make a service request and return the result.
