+++ maliput_ros/test/maliput_ros/ros/CMakeLists.txt
@@ -59,8 +59,13 @@ install(
   RUNTIME DESTINATION ${TEST_PLUGIN_INSTALL_DIR}
 )
 
+ament_add_gmock(branch_point_service_test branch_point_service_test.cc)
+ament_add_gmock(junction_service_test junction_service_test.cc)
+ament_add_gmock(lane_service_test lane_service_test.cc)
 ament_add_gmock(maliput_query_test maliput_query_test.cc)
 ament_add_gmock(maliput_query_node_test maliput_query_node_test.cc)
+ament_add_gmock(road_geometry_service_test road_geometry_service_test.cc)
+ament_add_gmock(segment_service_test segment_service_test.cc)
 
 macro(add_dependencies_to_test target)
     if (TARGET ${target})
@@ -95,5 +100,10 @@ macro(add_dependencies_to_test target)
     endif()
 endmacro()
 
+add_dependencies_to_test(branch_point_service_test)
+add_dependencies_to_test(junction_service_test)
+add_dependencies_to_test(lane_service_test)
 add_dependencies_to_test(maliput_query_test)
 add_dependencies_to_test(maliput_query_node_test)
+add_dependencies_to_test(road_geometry_service_test)
+add_dependencies_to_test(segment_service_test)
